ErrorReportContainer
Does a rename, so that the old component hierarchy:

```
ErrorList > ErrorSublist > Error
```

Is now less goofily:

```
ErrorReport > ErrorList > Error
```

Also creates an `ErrorReport` container which connects the error tree to
the Redux store.
