+++ src/components/index.js
@@ -1,5 +1,7 @@
+import ErrorReport from './ErrorReport';
 import Preview from './Preview';
 
 export {
+  ErrorReport,
   Preview,
 };
