+++ src/containers/index.js
@@ -1,5 +1,7 @@
 import Preview from './Preview';
+import ErrorReport from './ErrorReport';
 
 export {
+  ErrorReport,
   Preview,
 };
